% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_functions.R
\name{fava_norm}
\alias{fava_norm}
\title{Compute the normalized Fst of a matrix of compositional vectors}
\usage{
fava_norm(Q, K = ncol(Q))
}
\arguments{
\item{Q}{A matrix with \code{I=nrow(Q)} rows, each containing \code{K=ncol(Q)} non-negative entries that sum to 1.
If \code{Q} contains any metadata, it must be on the left-hand side of the matrix and the number of entries
that sum to 1 (\code{K}) must be specified.}

\item{K}{Optional; an integer specifying the number of categories in the data. Default is \code{K=ncol(Q)}.}
}
\value{
A numeric value between 0 and 1.
}
\description{
This function computes the normalized Fst given the number of rows and the mean abundance of the most abundant category.
We employ the normalization employed in the \href{https://github.com/MaikeMorrison/FSTruct}{FSTruct package} by
\href{https://doi.org/10.1111/1755-0998.13647}{Morrison, Alcala, and Rosenberg (2020)}.
}
\examples{
# Compute the weighted fava of
# the following compositional vectors:
q1 = c(1,   0,   0,   0)
q2 = c(0.5, 0.5, 0,   0)
q3 = c(1/4, 1/4, 1/4, 1/4)
q4 = c(0,   0,   1,   0)
Q_matrix = matrix(c(q1, q2, q3, q4),
                  byrow = TRUE, nrow = 4)

fava_norm(Q_matrix)
}
