% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_functions.R
\name{hetMean}
\alias{hetMean}
\title{Compute the mean heterozygosity of the rows in a matrix of compositional vectors}
\usage{
hetMean(Q, w = rep(1/nrow(Q), nrow(Q)), S = diag(ncol(Q)))
}
\arguments{
\item{Q}{A matrix with \code{I=nrow(Q)} rows, each containing \code{K=ncol(Q)} non-negative entries that sum to 1.}

\item{w}{Optional; a vector of length \code{I} with non-negative entries that sum to 1. Entry \code{w[i]} represents the weight placed on row \code{i} in the computation of the mean variability across rows. The default value is \code{w = rep(1/nrow(Q), nrow(Q))}.}

\item{S}{Optional; a K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry \code{S[i,k]} for \code{i!=k} is the similarity between category and \code{i} and category \code{k}, equalling 1 if the categories are to be treated as identical and equaling 0 if they are to be treated as totally dissimilar. The default value is \code{S = diag(ncol(Q))}.}
}
\value{
A numeric value between 0 and 1.
}
\description{
This function computes the mean heterozygosity, a statistical measure of variability also known as the Gini-Simpson index, of a set of vectors of non-negative entries which sum to 1. The function returns a number between 0 and 1 which quantifies the mean variability of the vectors. Values of 0 are achieved when each vector is a permutation of (1,0,..., 0). The value approaches 1 as the number of categories K increases when the vectors are equal to (1/K, 1/K, ..., 1/K).
}
\examples{
# To compute the mean heterozygosity of
# the following compositional vectors...
q1 = c(1,   0,   0,   0)
q2 = c(0.5, 0.5, 0,   0)
q3 = c(1/4, 1/4, 1/4, 1/4)
q4 = c(0,   0,   1,   0)

# we could compute the mean manually:
mean(sapply(list(q1, q2, q3, q4), het))

# Or we could use hetMean:
Q_matrix = matrix(c(q1, q2, q3, q4),
                  byrow = TRUE, nrow = 4)

hetMean(Q_matrix)

# Incoporating weights:

# Compute mean heterozygosity ignoring
# rows 2 and 3
row_weights = c(0.5, 0, 0, 0.5)
hetMean(Q_matrix, w = row_weights)

# Compute mean heterozygosity assuming that
# categories 1 and 2 are identical:
similarity_matrix = diag(4)
similarity_matrix[1,2] = 1
similarity_matrix[2,1] = 1
hetMean(Q_matrix, S = similarity_matrix)

# Assume categories 1 and 2 are identical AND
# ignore rows 2 and 4:
row_weights = c(0.5, 0, 0.5, 0)
hetMean(Q_matrix, w = row_weights, S = similarity_matrix)
}
