% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_functions.R
\name{hetPooled}
\alias{hetPooled}
\title{Compute the pooled heterozygosity of the rows in a matrix of compositional vectors}
\usage{
hetPooled(Q, K = ncol(Q), w, S)
}
\arguments{
\item{Q}{A matrix with \code{I=nrow(Q)} rows, each containing \code{K=ncol(Q)} non-negative entries that sum to 1.}

\item{K}{Optional; an integer specifying the number of categories in the data. Default is \code{K=ncol(Q)}.}

\item{w}{Optional; a vector of length \code{I} with non-negative entries that sum to 1. Entry \code{w[i]} represents the weight placed on row \code{i} in the computation of the mean abundance of each category across rows. The default value is \code{w = rep(1/nrow(Q), nrow(Q))}.}

\item{S}{Optional; a K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry \code{S[i,k]} for \code{i!=k} is the similarity between category and \code{i} and category \code{k}, equalling 1 if the categories are to be treated as identical and equaling 0 if they are to be treated as totally dissimilar. The default value is \code{S = diag(ncol(Q))}.}
}
\value{
A numeric value between 0 and 1.
}
\description{
This function computes the heterozygosity of a "pooled" vector equal to \code{colMeans(Q)}. Values of 0 are achieved when this pooled vector is a permutation of (1,0,..., 0). The value approaches 1 as the number of categories K increases when this pooled vector is equal to (1/K, 1/K, ..., 1/K).
}
\examples{
# To compute the pooled heterozygosity of
# the following compositional vectors...
q1 = c(1,   0,   0,   0)
q2 = c(0.5, 0.5, 0,   0)
q3 = c(1/4, 1/4, 1/4, 1/4)
q4 = c(0,   0,   1,   0)

# we could compute the mean manually:
qPooled = (q1 + q2 + q3 + q4)/4
het(qPooled)

# Or we could use hetPooled:
Q_matrix = matrix(c(q1, q2, q3, q4),
                  byrow = TRUE, nrow = 4)

hetPooled(Q_matrix)

# Incoporating weights:

# Compute pooled heterozygosity ignoring
# rows 2 and 3
row_weights = c(0.5, 0, 0, 0.5)
hetPooled(Q_matrix, w = row_weights)

# Compute pooled heterozygosity assuming that
# categories 1 and 2 are identical:
similarity_matrix = diag(4)
similarity_matrix[1,2] = 1
similarity_matrix[2,1] = 1
hetPooled(Q_matrix, S = similarity_matrix)

# Assume categories 1 and 2 are identical AND
# ignore rows 2 and 4:
row_weights = c(0.5, 0, 0.5, 0)
hetPooled(Q_matrix, w = row_weights, S = similarity_matrix)
}
