% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{Q_plot}
\alias{Q_plot}
\title{Plot a compositional matrix using ggplot2}
\usage{
Q_plot(Q, K = ncol(Q), arrange, group)
}
\arguments{
\item{Q}{A dataframe, matrix, or array representing compositional data. While each row
represents a compositional vector, if \code{K} is specified the first
\code{ncol(Q)-K} columns may contain other information. When restricted to the
last \code{K} columns, the rows of this matrix must sum to 1.}

\item{K}{Optional; the number of categories in the Q matrix. Each vector must
must have \code{K} categories. Default is the number of columns in \code{Q}.}

\item{arrange}{Optional; controls horizontal ordering of samples and vertical ordering of categories.
If \code{arrange = TRUE} or \code{arrange = "both"}, samples are ordered by the categories of greatest
abundance and categories are ordered in decreasing abundance. If \code{arrange = "vertical"}, sample
order is unchanged but categories are ordered in decreasing abundance. If \code{arrange = "horizontal"},
samples are ordered by the most abundant categories, but category order is unchanged. If \code{arrange} is missing
or \code{arrange = FALSE}, neither order is changed.}

\item{group}{Optional; a string specifying the name of the column that describes which group each row  belongs to. Use if \code{Q} is a single matrix containing multiple groups of individuals you wish to compare.}
}
\value{
A ggplot object containing a bar plot visualization of the Q matrix.
}
\description{
This function enables graphical visualization of a matrix of compostional data.
In the output plot, each vertical bar represents a single vector;
the height of each color in the bar corresponds to the abundance of each category
in that vector. Because this function produces a
ggplot object, its output can be modified using
standard ggplot2 syntax.
}
\examples{
Q_plot(
  # Make an example matrix of compositional data
  # Each row is an individual. Rows sum to 1.
  Q = matrix(c(
    .4, .2, .4,
    .5, .3, .2,
    .5, .4, .1,
    .6, .1, .3,
    .6, .3, .1
  ),
  nrow = 5,
  byrow = TRUE
  ),
 K = 3, # How many categories per vector?
 arrange = TRUE
) +
  # Below are example, optional modifications to the default plot
  ggplot2::ggtitle("Population A") +
  ggplot2::scale_fill_brewer("Blues") +
  ggplot2::scale_color_brewer("Blues") +
  ggplot2::xlab("Individuals")
  # Note that both scale_fill and scale_color are needed to change the color of the bars.
}
