% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{relab_phyloseq}
\alias{relab_phyloseq}
\title{Generate a relative abundance matrix with sample meta data and OTU abundances from a  phyloseq object.}
\usage{
relab_phyloseq(phyloseq_object)
}
\arguments{
\item{phyloseq_object}{A phyloseq object containing both an OTU table (\code{otu_table}) and sample metadata (\code{sample_data}).}
}
\value{
A data frame with rows representing samples and columns representing sample data categories or OTU relative abundances.
}
\description{
The R package phyloseq streamlines the storage and analysis of microbiome sequence data. This function takes a phyloseq object and extracts the OTU table and the sample meta data and combines them into one relative abundance matrix with rows corresponding to samples, meta data on the left-hand side, and OTU relative abundances on the right-hand side.
}
\examples{
otu_table = relab_phyloseq(FAVA::xue_phyloseq)
otu_table[1:10, 1:6]
}
