% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{window_fava}
\alias{window_fava}
\title{Compute FAVA in sliding windows.}
\usage{
window_fava(
  Q,
  K = ncol(Q),
  window_size,
  group,
  window_step = 1,
  normalized = FALSE,
  w = rep(1/nrow(Q), nrow(Q)),
  S = diag(K),
  index
)
}
\arguments{
\item{Q}{A matrix or data frame with rows containing non-negative entries that sum to 1. Each row represents
a sample, each column represents a category, and each entry represents the abundance of that category in the sample.
If \code{Q} contains any metadata, it must be on the left-hand side of the matrix,
the right \code{K} entries of each row must sum to 1, and \code{K} must be specified. Otherwise, all entries of
each row must sum to 1.}

\item{K}{Optional; number of categories which sum to 1 for each sample. Default is \code{K=ncol(Q)}.}

\item{window_size}{An integer number specifying the number of samples per window.}

\item{group}{Optional; a string specifying the name of the column that describes which group each row (sample) belongs to. Use if \code{Q} is a
single matrix containing multiple groups of samples you wish to compare, such as samples from different subjects.}

\item{window_step}{Optional; an integer specifying the distance between the first entry of adjacent windows. Default is \code{window_step=1}.}

\item{normalized}{Default is \code{normalized = FALSE}. If \code{normalized = TRUE}, then Fst is normalized by the theoretical upper bound conditional
on the most abundant category and the number of categories. This setting is recommended when there are fewer than 5 samples per window. Note that it is
incompatible with weightings at this time.}

\item{w}{Optional; a vector of length \code{I} with non-negative entries that sum to 1. Entry \code{w[i]} represents the weight placed on row \code{i} i
n the computation of the mean abundance of each category across rows. The default value is \code{w = rep(1/nrow(Q), nrow(Q))}. If \code{group} is specified,
then \code{w} can contain the weights for multiple groups, concatenated in the same order as the rows of \code{Q}. In this case, \code{w} must sum to the
number of distinct elements of \code{group}.}

\item{S}{Optional; a K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry \code{S[i,k]} for
\code{i!=k} is the similarity between category and \code{i} and category \code{k}, equalling 1 if the categories are to be treated as identical and equaling
0 if they are to be treated as totally dissimilar. The default value is \code{S = diag(ncol(Q))}.}

\item{index}{Optional; the name of the column in \code{Q} containing an index for each sample. For example, if \code{Q} contains time series data,
\code{index} would be the column containing the time of each sample.}
}
\value{
A list of values of Fst for each window.
}
\description{
This function computes FAVA in sliding window slices of a dataset.
}
\examples{
A = matrix(c(.3,.7,0,.1,0,.9,.2,.5,.3,.1,.8,.1,.3,.4,.3,.6,.4,0,0,.5,.5),
           ncol = 3, byrow = TRUE)
window_fava(Q = A, window_size = 4, normalized = TRUE)
}
