% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{xue_microbiome_sample}
\alias{xue_microbiome_sample}
\title{Temporal microbiome composition data}
\format{
\subsection{\code{xue_microbiome_sample}}{

A data frame with 75 rows and 1,348 columns:
\describe{
\item{subject}{Subject ID: XBA, XDA, or XMA}
\item{timepoint}{Time (days) of sample collection}
\item{...}{Species names}
...
}
}
}
\usage{
xue_microbiome_sample
}
\description{
A subset of the data generated by Xue et al. (2024) detailing longitudinal
composition of the human gut microbiome for three subjects who experience an
antiobitic perturbation between days 29 and 34. We include only the subjects
XAA, XBA, and XCA. Data from \doi{10.1101/2023.09.26.559480}.
}
\keyword{datasets}
