% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{xue_species_info}
\alias{xue_species_info}
\title{Table of species information}
\format{
\subsection{\code{xue_species_info}}{

A data frame with 1346 rows and 9 columns:
\describe{
\item{species_id}{The species_id given in xue_microbiome_sample}
\item{kingdom, phylum, class, order, family, genus, species}{The corresponding taxonomic category for each species}
\item{species_id_number}{The numeric code associated with each species, as used in xue_species_tree}
}
}
}
\usage{
xue_species_info
}
\description{
A data frame providing taxonomic information for the species included in xue_species_tree.
}
\keyword{datasets}
