% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_fava.R
\name{bootstrap_fava}
\alias{bootstrap_fava}
\title{Generate and analyze bootstrap replicates of one or more relative abundance matrices}
\usage{
bootstrap_fava(
  matrices,
  n_replicates,
  seed = NULL,
  group = NULL,
  time = NULL,
  w = NULL,
  K = NULL,
  S = NULL,
  normalized = FALSE,
  save_replicates = TRUE
)
}
\arguments{
\item{matrices}{A dataframe, matrix, or array representing a relative abundance matrix or a (possibly named) list of arbitrarily many relative abundance matrices. For each relative abundance matrix, matrix rows represent samples and the last \code{K} columns contain relative abundances of distinct categories (when restricted to the last \code{K} columns, the rows must sum to 1). If the matrices are not named (e.g., \code{matrices = list(matrix1, matrix2)} instead of \code{matrices = list(A = matrix1, B = matrix2)}), the matrices will be numbered in the order they are provided in the list. If \code{matrices} is a single matrix, dataframe, or array and \code{group} is specified, the matrix will be split into multiple relative abundance matrices, one for each distinct value of the column \code{group}, which will each be analyzed separately.}

\item{n_replicates}{The number of bootstrap replicate matrices to generate for each provided relative abundance matrix.}

\item{seed}{Optional; a number to set as the random seed. Use if reproducibility of random results is desired.}

\item{group}{Optional; a string specifying the name of the column that describes which group each row (sample) belongs to. Use if \code{matrices} is a single matrix containing multiple groups of samples you wish to compare.}

\item{time}{Optional; a string specifying the name of the column that describes the sampling time for each row. Include if you wish to weight FAVA by the distance between samples.}

\item{w}{Optional; a vector of length \code{nrow(matrices)} (if \code{matrices} is a dataframe, matrix, or array and \code{group} is not provided) or a list of vectors, each corresponding to one group (if \code{matrices} is a dataframe, matrix, or array and \code{group} is provided) or one matrix (if \code{matrices} is a list of matrices). Each vector must contain non-negative entries that sum to 1. Entry \code{w[i]} represents the weight placed on row \code{i} in the computation of the mean abundance of each category across rows. The default value is \code{w = rep(1/I, nrow(relab_matrix))} where \code{I} is the number of rows in each matrix/group.}

\item{K}{Optional; the number of categories in each provided relative abundance matrix, or a vector of such K values if the number of categories differs between matrices. If a single K is provided, each sample in every matrix must have \code{K} categories. If a vector of multiple K values is provided, \code{matrices} must be a list and the \eqn{i^{th}} entry of \code{K} must correspond to the \eqn{i^{th}} Q matrix in \code{matrices}. The default value of \code{K} is the number of columns in the matrix, the number of columns in the first matrix if a list is provided, or the number of columns minus 1 if \code{group} is specified but \code{K} is not.}

\item{S}{Optional; a K x K similarity matrix with diagonal elements equal to 1 and off-diagonal elements between 0 and 1. Entry \code{S[i,k]} for \code{i!=k} is the similarity between category and \code{i} and category \code{k}, equalling 1 if the categories are to be treated as identical and equaling 0 if they are to be treated as totally dissimilar. The default value is \code{S = diag(ncol(relab_matrix))}.}

\item{normalized}{Optional; should normalized FAVA be used? Default is \code{normalized = FALSE}; use \code{normalized = TRUE} to compute normalized FAVA. FAVA can only be normalized if it is not weighted.}

\item{save_replicates}{Optional; should all of the bootstrap replicate matrices be included in the output? Default is \code{save_replicates = FALSE}; use \code{save_replicates = FALSE} to save memory when analyzing large datasets.}
}
\value{
A named list containing the following entries:
\itemize{
\item \code{bootstrap_replicates}: A named list of lists. Each element is named for a relative abundance matrix provided in \code{matrices} and contains a list of \code{n_replicates} bootstrap replicates of the provided matrix. E.g., if \code{n_replicates = 100} and the first relative abundance matrix in \code{matrices} is named \code{A}, then the first element of \code{bootstrap_replicates}, \code{bootstrap_replicates$bootstrap_matrices_A}, is itself a list of 100 matrices, each representing a bootstrap replicate of matrix A.
\item \code{statistics}: A dataframe containing the desired statistic (FAVA, normalized FAVA, or weighted FAVA), computed for each bootstrap replicate matrix in \code{bootstrap_replicates}. The first column, titled \code{Matrix}, is a factor indicating which provided relative abundance matrix the row corresponds to (the matrix name if \code{matrices} is a named list, or a number otherwise). The row names are of the form \code{stats_matrix.replicate} where \code{matrix} is the name of one of the provided relative abundance matrices (or the entry number if the list elements were not named) and replicate is the number of bootstrap replicate (rep takes values from 1 to \code{n_replicates}).
\item \code{plot_boxplot}: A ggplot2 box plot depicting the bootstrap distribution of FAVA for each matrix in \code{matrices}.
\item \code{plot_violin}: A ggplot2 violin plot depicting the bootstrap distribution of FAVA for each matrix in \code{matrices}.
\item \code{plot_ecdf}: A ggplot2 empirical cumulative distribution function plot depicting the bootstrap distribution of FAVA for each matrix in \code{matrices}.
\item \code{test_kruskal_wallis}: Results of a Kruskal-Wallis test performed on the bootstrap distributions of FAVA. This test is a non-parametric statistical test of whether all provided bootstrap distributions are identically distributed.
\item \code{test_pairwise_wilcox}: Results of a Wilcoxon rank-sum test performed on the bootstrap distributions of FAVA. This test is a non-parameteric statistical test of whether \emph{each pairwise combination} of provided bootstrap distributions is identically distributed. The result is a matrix of p-values whose entries correspond to each pair of relative abundance matrices.
}
}
\description{
Generates bootstrap replicate relative abundance matrices, computes fava (possibly weighted or normalized) for each bootstrap replicate, produces several plots of the bootstrap distributions of Fst (possibly weighted) and/or Fst/FstMax for each provided Q matrix, and runs two statistical tests comparing these bootstrap distributions. The tests comparing bootstrap distributions of Fst (possibly weighted) and/or Fst/FstMax facilitate statistical comparison of the variability in each of multiple Q matrices.
}
\examples{

# Generate 100 bootstrap replicates for each of the
# 3 subjects included in the example data, xue_microbiome_sample.
# In this example, we:
# use the group argument so that each subject is analyzed,
# provide a species similarity matrix to incorporate phylogenetic information,
# and use the time argument so that samples are weighted by the amount of
# time between the samples before and after it.
boot_out = bootstrap_fava(matrices = xue_microbiome_sample,
               n_replicates = 100,
               seed = 1,
               group = "subject",
               time = "timepoint",
               S = xue_species_similarity,
               save_replicates = TRUE)

# To look at a plot of the distribution of
# FAVA for each Q matrix:
boot_out$plot_violin

# To determine if each of the 4 distibutions of
# FAVA is significantly different from
# each of the other distributions:
boot_out$test_pairwise_wilcox



}
